Threads user and policy dependencies to specific submodule resources
The module-level depends_on causes all the data sources to be read
during apply, when the module-dependency changes. This causes a lot
of noise in the plan, and makes it appear as if things are changing
when they really are not. By passing the depends_on values to specific
resources in the submodules, terraform properly maps the dependency
tree and generates a more meaningful plan.