Continues to use map(string) for top-level inline_policy template_vars
For an as-yet unknown reason, using `any` for the template_vars type
in the top-level inline_polcies variables causes terraform to panic
with the error:

```
panic: inconsistent list element types
```

No workaround become evident, so for now this will be a limitation
when invoking the top-level module and passing inline_policies with
template_vars. A workaround is to use each nested module independently,
which do work when the type is `any`.