Flattens the policy document objects to accept pure strings
This will require that users separately resolve the policy_document,
any way they like. They can use the included submodule, policy_document,
which accepts the same template, template_paths, and template_vars
arguments that were used previously.

The intent here is to update the policy_document module to support
`any` as the template_var type, which is supported by the templatefile()
function. However, there is a limitation in terraform's collections,
while you can have a list(any) for example, and it will indeed accept
any type, all objects in the list still must be the same type. Kind of
defeats the purpose IMO, but it is what it is.

The prior approach ran afoul of this limitation, when the user provided
template_vars per policy_document, and different template_vars objects
had different types. Which is not unlikely because different policy_documents
would map to different policies, and have different policy_template
variables. Policy templates might use jsonencode with complex objects,
which works fine with templatefile(). But this fails when providing lists
of template_vars where each item may have different types.

So instead, we're now asking the user to resolve the policy_documents
separately. That will allow them to resolve complex types in the templates,
in any way they like that will actually work within terraform's constraints.